// src/module/tours/dto/update-activity.dto.ts
import { ActivityStatus } from '@prisma/client';
export class UpdateActivityDto {
  supplierId?: number;
  destinationId?: number;
  categoryId?: number;
  name?: string;
  slug?: string;
  description?: string;
  highlights?: any;
  duration?: number;
  price?: number;
  currency?: string;
  maxParticipants?: number;
  rating?: number;
  reviewCount?: number;
  instantConfirmation?: boolean;
  freeCancellation?: boolean;
  status?: ActivityStatus;
  featured?: boolean;
}